# packages/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DEMS API",
    "description": "Plataforma de evaluación de prototipos IPN-DEMS.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local dev"
    }
  ],
  "paths": {
...
    "/admin/results/categoria/{slug}": {
      "get": {
        "tags": [
          "admin/results"
        ],
        "operationId": "by_categoria",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "Slug de la categoría",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "edition_id",
            "in": "query",
            "description": "ID de la edición",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "rubric_type",
            "in": "query",
            "description": "exhibicion (default) | memoria",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ranking de prototipos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoriaResultsView"
                }
              }
            }
          },
          "400": {
            "description": "rubric_type inválido o edition_id ausente"
          },
          "403": {
            "description": "Sólo admin"
          },
          "404": {
            "description": "Categoría o edición no encontrada"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/admin/results/edition/{id}/export.csv": {
      "get": {
        "tags": [
          "admin/results"
        ],
        "summary": "CSV con todos los prototipos de la edición. Una fila por (categoría,\nprototipo). Promedio sobre evaluaciones submitted del rubric_type\nsolicitado.",
        "operationId": "export_csv",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID de la edición",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "rubric_type",
            "in": "query",
            "description": "exhibicion (default) | memoria",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CSV de resultados",
            "content": {
              "text/csv": {}
            }
          },
          "400": {
            "description": "rubric_type inválido"
          },
          "403": {
            "description": "Sólo admin"
          },
          "404": {
            "description": "Edición no encontrada"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/admin/results/edition/{id}/final": {
      "get": {
        "tags": [
          "admin/results"
        ],
        "summary": "Puntaje final combinado por prototipo de una edición.",
        "description": "Para cada prototipo: `puntaje_final = Σ_tipo (promedio_tipo / max_total_tipo)\n* (peso_tipo / 100)`, considerando SÓLO la rúbrica ACTIVA de cada tipo. El\npromedio_tipo es la media de los `total` (suma de scores) de las\nevaluaciones ya `submitted_at` contra esa rúbrica. Si un tipo no tiene\nrúbrica activa o `max_total = 0`, su aporte es 0 (no NaN). El ranking se\nordena por `puntaje_final` descendente.",
        "operationId": "final_ranking",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID de la edición",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ranking combinado ponderado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinalRankingView"
                }
              }
            }
          },
          "403": {
            "description": "Sólo admin"
          },
          "404": {
            "description": "Edición no encontrada"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CategoriaRef": {
        "type": "object",
        "required": [
          "id",
          "slug",
          "nombre"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nombre": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        }
      },
      "CategoriaResultsView": {
        "type": "object",
        "required": [
          "categoria",
          "edition_id",
          "rubric_type",
          "max_total",
          "prototipos"
        ],
        "properties": {
          "categoria": {
            "$ref": "#/components/schemas/CategoriaRef"
          },
          "edition_id": {
            "type": "string",
            "format": "uuid"
          },
          "max_total": {
            "type": "integer",
            "format": "int64",
            "description": "Suma máxima alcanzable según la rúbrica activa (0 si no hay rúbrica)."
          },
          "prototipos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PrototipoResultView"
            }
          },
          "rubric_type": {
            "type": "string"
          }
        }
      },
...
      "EvaluacionResultView": {
        "type": "object",
        "required": [
          "evaluacion_id",
          "jurado_id",
          "jurado_nombre",
          "total",
          "submitted_at"
        ],
        "properties": {
          "evaluacion_id": {
            "type": "string",
            "format": "uuid"
          },
          "jurado_id": {
            "type": "string",
            "format": "uuid"
          },
          "jurado_nombre": {
            "type": "string"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "FinalBreakdownView": {
        "type": "object",
        "required": [
          "rubric_type",
          "peso",
          "max_total",
          "aporte"
        ],
        "properties": {
          "aporte": {
            "type": "number",
            "format": "double",
            "description": "Aporte ponderado de este tipo al `puntaje_final`:\n(promedio / max_total) * (peso / 100). 0 si max_total es 0."
          },
          "max_total": {
            "type": "integer",
            "format": "int64"
          },
          "peso": {
            "type": "integer",
            "format": "int32"
          },
          "promedio": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Promedio de los `total` de jurados (null si no hay evaluaciones)."
          },
          "rubric_type": {
            "type": "string"
          }
        }
      },
      "FinalPrototipoView": {
        "type": "object",
        "required": [
          "prototipo_id",
          "folio",
          "nombre",
          "puntaje_final",
          "desglose"
        ],
        "properties": {
          "desglose": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FinalBreakdownView"
            },
            "description": "Desglose por tipo de rúbrica (exhibicion, memoria)."
          },
          "folio": {
            "type": "string"
          },
          "nombre": {
            "type": "string"
          },
          "prototipo_id": {
            "type": "string",
            "format": "uuid"
          },
          "puntaje_final": {
            "type": "number",
            "format": "double",
            "description": "Σ_tipo (promedio_tipo / max_total_tipo) * (peso_tipo / 100), en 0..1."
          }
        }
      },
      "FinalRankingView": {
        "type": "object",
        "required": [
          "edition_id",
          "prototipos"
        ],
        "properties": {
          "edition_id": {
            "type": "string",
            "format": "uuid"
          },
          "prototipos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FinalPrototipoView"
            },
            "description": "Prototipos ordenados por `puntaje_final` descendente."
          }
        }
      },
...
      "PrototipoResultView": {
        "type": "object",
        "required": [
          "prototipo_id",
          "folio",
          "nombre",
          "n_jurados",
          "evaluaciones"
        ],
        "properties": {
          "evaluaciones": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EvaluacionResultView"
            }
          },
          "folio": {
            "type": "string"
          },
          "n_jurados": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          },
          "promedio": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Promedio de los `total` por jurado (null si no hay evaluaciones)."
          },
          "prototipo_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}